+++ nodes/admin.py
@@ -6,4 +6,3 @@ admin.site.register(Subject)
 admin.site.register(Module)
 admin.site.register(Unit)
 admin.site.register(Lesson)
-admin.site.register(LessonResult)
