+++ tasks/models.py
@@ -1,6 +1,7 @@
 from django.db import models
 from users.models import User
-from nodes.models import Lesson, LessonResult
+from nodes.models import Lesson
+from results.models import LessonResult
 
 
 # Tasks
@@ -22,7 +23,7 @@ class LessonTask(Task):
         verbose_name_plural = 'Домашние задания, уроки'
 
     lesson = models.ForeignKey(Lesson)
-    lesson_result = models.ForeignKey(LessonResult, blank=True, null=True)
+    # lesson_result = models.ForeignKey(LessonResult, blank=True, null=True)
 
     def __str__(self):
         return u'For {}, "{}"'.format(self.student, self.lesson)
